+++ src/modules/pages/page-editor/launch-dialog/index.tsx
@@ -13,10 +13,7 @@ import strings from "../../../../localization/index";
 import { DescriptionText } from "./styles";
 import theme from "../../../../theme";
 import { useDispatch } from "react-redux";
-import {
-  ComponentType,
-  IUserComponent,
-} from "../../../../store/user-pages/types";
+import { ButtonType, IUserComponent } from "../../../../store/user-pages/types";
 import { addMiddleComponentInPage } from "../../../../store/user-pages/actions";
 import MobileDatePicker from "@mui/lab/MobileDatePicker";
 import TimePicker from "@mui/lab/TimePicker";
@@ -87,7 +84,7 @@ const LaunchDialog = ({ pageId, open, handleClose }: IIconsDialogProps) => {
         columns: 2,
       },
       launchDate: launch.toISOString(),
-      type: ComponentType.Launch,
+      type: ButtonType.Launch,
       mediaUrl: message,
       iconDetails: undefined,
     };