+++ changelog.md
@@ -16,6 +16,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Versions
 
+### [1.19.0] - 2022-10-27
+
+- Feature: Removed the deleteImage request from everywhere, but the gallery.
+- Feature: Implemented animation of scale when hovering page renderer components
+
 ### [1.18.0] - 2022-10-26
 
 - Feature: added shadow feature in the button creation